+++ src/main/java/io/supertokens/webserver/api/saml/CreateSamlLoginRedirectAPI.java
@@ -23,6 +23,7 @@ import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.multitenancy.exceptions.TenantOrAppNotFoundException;
 import io.supertokens.saml.SAML;
 import io.supertokens.saml.exceptions.InvalidClientException;
+import io.supertokens.saml.exceptions.SAMLRequestSigningException;
 import io.supertokens.webserver.InputParser;
 import io.supertokens.webserver.WebserverAPI;
 import jakarta.servlet.ServletException;
@@ -69,7 +70,7 @@ public class CreateSamlLoginRedirectAPI extends WebserverAPI {
             res.addProperty("status", "INVALID_CLIENT_ERROR");
             super.sendJsonResponse(200, res, resp);
         } catch (TenantOrAppNotFoundException | StorageQueryException | CertificateEncodingException |
-                 FeatureNotEnabledException e) {
+                 FeatureNotEnabledException | SAMLRequestSigningException e) {
             throw new ServletException(e);
         }
     }